Move change log tests from StorageTest to ChangeLogTest
